add bhs

+++ app/src/main/java/com/febrian/covidapp/global/GlobalResponse.kt
@@ -6,5 +6,5 @@ data class GlobalResponse(
     val cases: Int?,
     val recovered: Int?,
     val deaths: Int?,
-    val updated : BigInteger?
+    val updated : Long?
 )
